pyproject.toml
[project]
name = "os-control"
version = "0.1.0"
description = "MCP servers for controlling macOS and Windows"
requires-python = ">=3.8"
dependencies = [
    "mcp>=0.1.0",
    "pydantic>=2.0.0",
    "watchdog>=3.0.0"
]

# Optional dependencies based on which servers you use
[project.optional-dependencies]
calendar = [
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.0.0",
    "pytz>=2024.1"
]

[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mac_control", "src/win_control", "src/file_watcher"]

[project.scripts]
mcp-server-mac = "mac_control.server:main"
mcp-server-win = "win_control.server:main"
mcp-server-files = "file_watcher.server:main" 